Add support for sending a cmd to a spez. channel
With this funktion, e. g. the hubot can send a command like "/echo hioh" to a channel.
in the coustom hubot script (.coffee):
msg.robot.adapter.client.postCommand('the_channel_id','/echo hioh');

In the future it would be nice, if 'the_channel_id' could either be the channel_id or the name of the channel. ('~channel_name')